Changes after feedback
